+++ tests/testthat/test-plot-print.R
@@ -13,6 +13,7 @@ if (skip_on_cran()) {
   segments <- fluss_extract(cac, nseg)
   chain <- find_chains(mp)
   motif <- find_motif(mp)
+  mps <- salient_subsequences(mp, n_bits = c(4, 6, 8), verbose = 0)
 
   plot_arcs_test <- function() plot_arcs(pairs = matrix(c(5, 10, 1, 10, 20, 5), ncol = 2, byrow = TRUE))
   plot_arccount_test <- function() plot.ArcCount(cac)
@@ -20,14 +21,17 @@ if (skip_on_cran()) {
   plot_fluss_test <- function() plot.Fluss(segments)
   plot_chain_test <- function() plot.Chain(chain)
   plot_motif_test <- function() plot.Motif(motif)
+  plot_salient_test <- function() plot.Salient(mps)
 
   mdata <- mp_toy_data$data[1:200, ]
   mw <- mp_toy_data$sub_len
   mmp <- tsmp(mdata, window_size = mw, mode = "mstomp", verbose = 0)
+  smp <- tsmp(mdata, window_size = mw, mode = "simple", verbose = 0)
   mmotif <- find_motif(mmp, n_motifs = 2)
 
   plot_multimatrixprofile_test <- function() plot.MultiMatrixProfile(mmp)
   plot_multimotif_test <- function() plot.MultiMotif(mmotif)
+  plot_simplematrixprofile_test <- function() plot.SimpleMatrixProfile(smp)
 
   test_that("Plot", {
     expect_doppelganger("plot arcs", plot_arcs_test)
@@ -38,6 +42,8 @@ if (skip_on_cran()) {
     expect_doppelganger("plot motif", plot_motif_test)
     expect_doppelganger("plot multi matrix profile", plot_multimatrixprofile_test)
     expect_doppelganger("plot multimotif", plot_multimotif_test)
+    expect_doppelganger("plot simple matrix profile", plot_simplematrixprofile_test)
+    expect_doppelganger("plot salient", plot_salient_test)
   })
 
   context("Testing Print")
@@ -53,8 +59,10 @@ if (skip_on_cran()) {
     expect_known_output(cac, file = paste0(path, "cac-print"), print = TRUE, update = upd)
     expect_known_output(mp, file = paste0(path, "mp-print"), print = TRUE, update = upd)
     expect_known_output(mmp, file = paste0(path, "mmp-print"), print = TRUE, update = upd)
+    expect_known_output(smp, file = paste0(path, "smp-print"), print = TRUE, update = upd)
     expect_known_output(chain, file = paste0(path, "chain-print"), print = TRUE, update = upd)
     expect_known_output(motif, file = paste0(path, "motif-print"), print = TRUE, update = upd)
     expect_known_output(mmotif, file = paste0(path, "mmotif-print"), print = TRUE, update = upd)
+    expect_known_output(mps, file = paste0(path, "salient-print"), print = TRUE, update = upd)
   })
 }